Add rewritten starClick() function.
Formerly, starClick referenced an interval in which the Observe Astronomical Events button was clicked. That clicker stopped working, but I got the new(er) selector so that I can use jQuery to click it. Instead of making it another interval, I made it into a function to match what I've done to the other functions. Thus, any time you call starClick(), it will click the Observe Astronomical Events button (assuming it is present).